requirements.txt
anyio==4.7.0
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
Django==5.1.4
graphql-core==3.2.5
h11==0.14.0
idna==3.10
libcst==1.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.43.0
strawberry-graphql==0.256.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
